C源码

+++ 第二章/100000568-《算法笔记》2.4小节——CC++快速入门-循环结构/5.1.9.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main(){
-	int x1=2,y1=1;
-	int i,x2=3,y2=2;
-	double sum=((double)x1)/y1+((double)x2)/y2;
+	double x1=2,y1=1;
+	double i,x2=3,y2=2;
+	double sum=x1/y1+x2/y2;
 	for(i=3;i<=20;i++){
-		int x3=x1+x2;
-		int y3=y1+y2;
-		double ele=((double)x3)/y3;
+		double x3=x1+x2;
+		double y3=y1+y2;
+		double ele=x3/y3;
 		sum+=ele;
 		x1=x2;
 		y1=y2;
